Make it a native-activity

+++ engine/common/drawing/font/font_manager.h
@@ -1,10 +1,12 @@
 /*
  * Copyright 2013 Jeremie Roy. All rights reserved.
- * License: http://www.opensource.org/licenses/BSD-2-Clause
+ * License: https://github.com/bkaradzic/bgfx#license-bsd-2-clause
  */
 
 #ifndef FONT_MANAGER_H_HEADER_GUARD
 #define FONT_MANAGER_H_HEADER_GUARD
+
+
 
 #include <bx/handlealloc.h>
 #include <bgfx/bgfx.h>
@@ -33,7 +35,7 @@ struct FontInfo
 	float descender;
 	/// The spacing in pixels between one row's descent and the next row's ascent.
 	float lineGap;
-	/// This field gives the maximum horizontal cursor advance for all glyphs in the font. 
+	/// This field gives the maximum horizontal cursor advance for all glyphs in the font.
 	float maxAdvanceWidth;
 	/// The thickness of the under/hover/strike-trough line in pixels.
 	float underlineThickness;
@@ -100,12 +102,12 @@ struct GlyphInfo
 	float offset_y;
 
 	/// For horizontal text layouts, this is the unscaled horizontal
-	/// distance in pixels used to increment the pen position when the 
+	/// distance in pixels used to increment the pen position when the
 	/// glyph is drawn as part of a string of text.
 	float advance_x;
 
 	/// For vertical text layouts, this is the unscaled vertical distance
-	/// in pixels used to increment the pen position when the glyph is 
+	/// in pixels used to increment the pen position when the glyph is
 	/// drawn as part of a string of text.
 	float advance_y;
 
@@ -125,7 +127,7 @@ public:
 
 	/// Create the font manager and create the texture cube as BGRA8 with
 	/// linear filtering.
-	FontManager(uint32_t _textureSideWidth = 512);
+	FontManager(uint16_t _textureSideWidth = 512);
 
 	~FontManager();
 
@@ -135,7 +137,7 @@ public:
 		return m_atlas;
 	}
 
-	/// Load a TrueType font from a given buffer. The buffer is copied and 
+	/// Load a TrueType font from a given buffer. The buffer is copied and
 	/// thus can be freed or reused after this call.
 	///
 	/// @return invalid handle if the loading fail
@@ -155,7 +157,7 @@ public:
 
 	/// Preload a set of glyphs from a TrueType file.
 	///
-	/// @return True if every glyph could be preloaded, false otherwise if 
+	/// @return True if every glyph could be preloaded, false otherwise if
 	///   the Font is a baked font, this only do validation on the characters.
 	bool preloadGlyph(FontHandle _handle, const wchar_t* _string);
 
@@ -167,7 +169,7 @@ public:
 	/// @remark the handle is required to be valid
 	const FontInfo& getFontInfo(FontHandle _handle) const;
 
-	/// Return the rendering informations about the glyph region. Load the 
+	/// Return the rendering informations about the glyph region. Load the
 	/// glyph from a TrueType font if possible
 	///
 	const GlyphInfo* getGlyphInfo(FontHandle _handle, CodePoint _codePoint);
